Moved admin tables to components ( users and products)
